move Trigger into Trigger.js
